[FE] fix: 플레이리스트 모달창 드롭다운버튼 컴포넌트 안으로 이동 #369
- 내 플레이리스트 추가 시에 뜨는 플레이리스트 모달창을 드롭다운버튼으로 이동
- 기존 트랙로우카드 컴포넌트에 있던 플레이리스트 모달컴포넌트 관련 코드 삭제
